feat(v8): Rename gcp package to `@sentry/google-cloud-serverless`
