+++ .github/ISSUE_TEMPLATE/bug.yml
@@ -39,11 +39,11 @@ body:
         - '@sentry/deno'
         - '@sentry/ember'
         - '@sentry/gatsby'
+        - '@sentry/google-cloud-serverless'
         - '@sentry/nextjs'
         - '@sentry/node'
         - '@sentry/react'
         - '@sentry/remix'
-        - '@sentry/google-cloud'
         - '@sentry/svelte'
         - '@sentry/sveltekit'
         - '@sentry/vue'
