+++ MIGRATION.md
@@ -171,12 +171,12 @@ The `Transaction` integration has been removed from `@sentry/integrations`. Ther
 #### @sentry/serverless
 
 `@sentry/serverless` has been removed and will no longer be published. The serverless package has been split into two
-different packages, `@sentry/aws-serverless` and `@sentry/google-cloud`. These new packages have smaller bundle size
-than `@sentry/serverless`, which should improve your serverless cold-start times.
+different packages, `@sentry/aws-serverless` and `@sentry/google-cloud-serverless`. These new packages have smaller
+bundle size than `@sentry/serverless`, which should improve your serverless cold-start times.
 
-`@sentry/aws-serverless` and `@sentry/google-cloud` has also been changed to only emit CJS builds. The ESM build for the
-`@sentry/serverless` package was always broken and we decided to remove it entirely. ESM support will be re-added at a
-later date.
+`@sentry/aws-serverless` and `@sentry/google-cloud-serverless` has also been changed to only emit CJS builds. The ESM
+build for the `@sentry/serverless` package was always broken and we decided to remove it entirely. ESM support will be
+re-added at a later date.
 
 In `@sentry/serverless` you had to use a namespace import to initialize the SDK. This has been removed so that you can
 directly import from the SDK instead.
@@ -209,7 +209,7 @@ Sentry.GCPFunction.init({
 });
 
 // v8
-const Sentry = require('@sentry/google-cloud');
+const Sentry = require('@sentry/google-cloud-serverless');
 
 Sentry.init({
   dsn: '__DSN__',
