+++ README.md
@@ -59,8 +59,8 @@ package. Please refer to the README and instructions of those SDKs for more deta
 - [`@sentry/remix`](https://github.com/getsentry/sentry-javascript/tree/master/packages/remix): SDK for Remix
 - [`@sentry/aws-serverless`](https://github.com/getsentry/sentry-javascript/tree/master/packages/aws-serverless): SDK
   for AWS Lambda Functions
-- [`@sentry/google-cloud`](https://github.com/getsentry/sentry-javascript/tree/master/packages/google-cloud): SDK for
-  Google Cloud Functions
+- [`@sentry/google-cloud-serverless`](https://github.com/getsentry/sentry-javascript/tree/master/packages/google-cloud):
+  SDK for Google Cloud Functions
 - [`@sentry/electron`](https://github.com/getsentry/sentry-electron): SDK for Electron with support for native crashes
 - [`@sentry/react-native`](https://github.com/getsentry/sentry-react-native): SDK for React Native with support for
   native crashes
