+++ packages/google-cloud-serverless/src/gcpfunction/general.ts
