+++ packages/google-cloud-serverless/test/__mocks__/dns.ts
